+++ lib/src/home/view/screens/android/submodules_android.dart
@@ -12,6 +12,9 @@ class SubmodulesListAndroid
     with ScrapbookFields {
   SubmodulesListAndroid() : super(I10n.s('Submodule'));
 
+  @override
+  SubmoduleFields get fields => SubmoduleFields();
+
   /// Flags indicating which fields are actually displayed.
   /// Note, flags are implemented in the mixin ScrapbookFields
   @override
@@ -41,18 +44,20 @@ class SubmodulesListAndroid
   @override
   bool useNextId = true;
 
-
   /// Supply the widgets to 'list' the record.
   @override
   List<Widget> addedWidgets(
-      Map<String, FieldWidgets<PlayHouseFields>> record,
-      VoidCallback onTap,
-      ) {
+    Map<String, FieldWidgets<PlayHouseFields>> record,
+    VoidCallback onTap,
+  ) {
+    final module = record['module_id'];
+    module.label = 'Module Id';
     final locked = record['lockedFirst'];
     locked.label = 'First Locked';
     final next = record['next_submodule_id'];
     next.label = 'Next Submodule';
     return [
+      module.onListTile(tap: onTap),
       locked.onListTile(tap: onTap),
       next.onListTile(tap: onTap),
     ];
@@ -61,37 +66,48 @@ class SubmodulesListAndroid
   /// Supply the fields required to 'edit' the current record.
   @override
   List<Widget> editWidgets(
-      Map<String, FieldWidgets<PlayHouseFields>> record,
-      ) {
+    Map<String, FieldWidgets<PlayHouseFields>> record,
+  ) {
     final moduleId = record['module_id'];
     moduleId.label = 'Module Id';
     final next = record['next_submodule_id'];
     return [
       record['rowid'].onListTile(enabled: false),
-      moduleId.onListItems(dropItems: _parentModule(moduleId)),
+      moduleId.onListItems(
+          onChanged: (String v) {
+            con.setState(() {
+              moduleId.value = int.parse(v);
+            });
+          },
+          dropItems: _parentModule(moduleId)),
       record['name'].textFormField,
       record['short_description'].textFormField,
       record['long_description'].textFormField,
-      next.onListItems(dropItems: _moduleItems(next)),
+      next.onListItems(
+          onChanged: (String v) {
+            con.setState(() {
+              next.value = int.parse(v);
+            });
+          },
+          dropItems: _moduleItems(record)),
     ];
   }
 
   /// List the parent modules
-  List<String> _parentModule(FieldWidgets<PlayHouseFields> record) {
-    return [''];
-  }
+  List<String> _parentModule(FieldWidgets<PlayHouseFields> record) =>
+      ModuleFields().table.idList;
 
   /// List the next submodules
-  List<String> _moduleItems(FieldWidgets<PlayHouseFields> record) {
-    return [''];
+  List<String> _moduleItems(Map<String, FieldWidgets<PlayHouseFields>> record) {
+    final rowid = record['rowid'].value.toString();
+    return fields.table.idList.where((id) => id != rowid).toList();
   }
-
 
   @override
   List<Map<String, FieldWidgets<PlayHouseFields>>> fetchData() =>
-      SubmoduleFields().field.values.toList();
+      fields.field.values.toList();
 
   @override
   Map<String, FieldWidgets<PlayHouseFields<SQLiteTable>>> newRecord() =>
-      SubmoduleFields().getNewRecord();
+      fields.getNewRecord();
 }